move the memory operation into another header
